requirements.txt
# Core ADK packages
google-adk==0.4.0
a2a-sdk==0.5.0

# LLM integration
litellm==1.74.0

# Core dependencies
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.26.0
joblib==1.3.2

# Machine Learning
scikit-learn==1.7.1

# Web server
uvicorn==0.28.0
starlette==0.35.1

# Utility packages
json5==0.9.14
requests==2.32.0
emailvalidator==1.0.5
